+++ include/main/planets/Gas.hpp
@@ -7,8 +7,8 @@ class Gas : public Planet {
 public:
   Gas(const Vector &_position, const Vector &_velocity);
   void drawPlanet() override;
-  int getSize();
+  __float128 getSize();
 
 private:
-  int size;
+  __float128 size;
 };
